#5172 newsletter refactoring: moves newsletter generation into its own NewsletterGenerator class
